handling some errors in %x

+++ ft_printf/ft_putnbr_base.c
@@ -6,7 +6,7 @@
 /*   By: oakhmouc <oakhmouc@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/07 14:32:49 by oakhmouc          #+#    #+#             */
-/*   Updated: 2024/12/09 18:48:40 by oakhmouc         ###   ########.fr       */
+/*   Updated: 2024/12/09 18:52:12 by oakhmouc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,7 +36,7 @@ int	base_len(char *base)
 	return (n);
 }
 
-int	ft_putnbr_base(int lnbr, char *base)
+int	ft_putnbr_base(long long int lnbr, char *base)
 {
 	int			lbase;
 	int			ret;
